Use BigDecimal for currency values

+++ app/src/main/java/com/augustg/shoppingcart/util/Formats.kt
@@ -1,5 +1,6 @@
 package com.augustg.shoppingcart.util
 
+import java.math.BigDecimal
 import java.text.NumberFormat
 import java.util.*
 
@@ -8,5 +9,5 @@ object Formats {
     fun getLocale(): Locale = locale
 
     private val currencyFormat = NumberFormat.getCurrencyInstance(locale)
-    fun getAsCurrency(amount: Double): String = currencyFormat.format(amount)
+    fun getAsCurrency(amount: BigDecimal): String = currencyFormat.format(amount)
 }
